+++ RELEASE_NOTES.md
@@ -3,6 +3,7 @@
 * Map XSD date to System.DateTime and XSD dateTime to System.DateTimeOffset
 * Fixed large float values being silently converted to int.MinValue when parsing JSON.
 * Improved handling of invalid cookies.
+* Fixes for #1091 - tidying up regex handling in script tags.
 
 #### 3.0.0-beta4 - July 8 2018
 * (Breaking Change) Ignore culture when parsing JSON to better match the JSON spec.